test: move tests to correct tests/ root per project convention
- Remove wrongly placed src/**/__tests__/ files and src/test-utils/with-setup.ts
- Add tests/settings-view.integration.test.ts
- Add tests/use-api-key-management.unit.test.ts
