+++ src/composables/__tests__/use-api-key-management.spec.ts
@@ -1,4 +1,1 @@
-// DELETED — replaced by use-api-key-management.unit.test.ts
-// This file is intentionally emptied; the old tests violated the testing guideline
-// (mock interaction assertions). See use-api-key-management.unit.test.ts.
 export {}
