+++ src/test-utils/with-setup.ts
@@ -1,2 +1,1 @@
-// DELETED — was only used by the removed bad tests.
 export {}
