+++ src/views/__tests__/SettingsView.spec.ts
@@ -1,4 +1,1 @@
-// DELETED — replaced by SettingsView.integration.test.ts
-// This file is intentionally emptied; the old tests violated the testing guideline
-// (mock interaction assertions, wrong layer). See SettingsView.integration.test.ts.
 export {}
